Fix typos in printargs family of programs
Additionally, removed an unecessary print from printargs1.c. It looks like you
may have started by separating the arguments with a single space, which is why
you needed the print at the end. But now it just leaves an extra blank line,
and is inconsistent with the output of printargs2 and printargs3.